allow for users to write :: instead of .[] because that gets irritating
The :: operator takes the value after the :: and converts that into a 
string. Then it wil convert that into a .[] statement. Because Omm has 
lazy indexes, you can do ::0 to get the first element of an array, even 
though ::0 becomes .['0']. if you want to get indexes by a variable, you 
can still use .[] notation.