fixed select() syscall
im using a pre increment trick now. before it was counting array indexes based on the "starting with index 1" model. i fixed that back to the standard starting with 0. using a pre increment trick B). (it was working on linux for some reason, but when i tested windows, i found this bug)